Move http utils to pkg
