--- internal/http/utils/utils.go
@@ -1,36 +1,0 @@
-// Copyright 2018-2020 CERN
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// In applying this license, CERN does not waive the privileges and immunities
-// granted to it by virtue of its status as an Intergovernmental Organization
-// or submit itself to any jurisdiction.
-
-package utils
-
-import (
-	"time"
-
-	types "github.com/cs3org/go-cs3apis/cs3/types/v1beta1"
-)
-
-// TSToUnixNano converts a protobuf Timestamp to uint64
-// with nanoseconds resolution.
-func TSToUnixNano(ts *types.Timestamp) uint64 {
-	return uint64(time.Unix(int64(ts.Seconds), int64(ts.Nanos)).UnixNano())
-}
-
-// TSToTime converts a protobuf Timestamp to Go's time.Time.
-func TSToTime(ts *types.Timestamp) time.Time {
-	return time.Unix(int64(ts.Seconds), int64(ts.Nanos))
-}
